src/handlers/evaluate_policy/requirements.txt
parliament==0.4.8
PyYAML==5.3
